DRIVERS-3473 Add client-side validation for maxAwaitTime+op timeout
